backend/requirements.txt
PyPDF2
python-docx
langchain
python-dotenv
openai
django
cloudinary
gunicorn
pillow
psycopg2-binary
requests
python-pptx
dj-database-url
djangorestframework 
djangorestframework-simplejwt
django-cloudinary-storage[video]
django-filter
django-cloudinary-storage
django-cors-headers==4.6.0
sqlparse==0.5.1
virtualenv==20.26.6
watchfiles
whitenoise
django-shortuuidfield
svix
clerk-backend-api
pydantic
fastapi
resend